R247_Dev_20210527/Designer/Python/segmentation/unet/requirement.txt
tensorflow-gpu== 2.5.0
numpy
easydict
tf2onnx
albumentations
omegaconf